move all validation code into validation package
